inner/requirements.txt
appnope==0.1.0
bleach==2.0.0
certifi==2017.4.17
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.3
grpcio==1.8.2
grpcio-tools==1.8.2
h5py==2.7.0
html5lib==0.999999999
Keras==2.0.2
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mock==2.0.0
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.0
pandas==0.20.2
pandocfilters==1.4.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.1.4
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
tensorflow==1.0.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
